Add CoreOperatingPointMovement test module
Move OperatingPointMovement tests to a Core-level test module so they
directly test CoreOperatingPointMovement behavior. Keep a slim
OperatingPointMovement test module that verifies the public subclass
relationship and inherited behavior. Duplicate the two
OperatingPointMovement fixtures that consumer tests need.
